+++ src/Sulu/Bundle/ContactBundle/Entity/Contact.php
@@ -12,6 +12,7 @@ namespace Sulu\Bundle\ContactBundle\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use JMS\Serializer\Annotation\Groups;
 use JMS\Serializer\Annotation\Accessor;
 use JMS\Serializer\Annotation\Exclude;
 use JMS\Serializer\Annotation\SerializedName;
@@ -83,11 +84,13 @@ class Contact extends ApiEntity implements ContactInterface, AuditableInterface
 
     /**
      * @var UserInterface
+     * @Groups({"fullContact"})
      */
     protected $changer;
 
     /**
      * @var UserInterface
+     * @Groups({"fullContact"})
      */
     protected $creator;